+++ packages/services/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @vegastack/pages-services
 
+## 0.1.14-next.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.1.14-next.8
+  - @vegastack/pages-db@0.1.14-next.8
+  - @vegastack/pages-renderer@0.1.14-next.8
+
+## 0.1.14-next.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.1.14-next.7
+  - @vegastack/pages-db@0.1.14-next.7
+  - @vegastack/pages-renderer@0.1.14-next.7
+
+## 0.1.14-next.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.1.14-next.6
+  - @vegastack/pages-db@0.1.14-next.6
+  - @vegastack/pages-renderer@0.1.14-next.6
+
 ## 0.1.14-next.5
 
 ### Patch Changes